refactor: clear page component (home)
