+++ src/components/global/CloseButton/styles.ts
