+++ src/components/global/Modal/ContentModal/styles.ts
