+++ src/components/global/Modal/styles.ts
