+++ src/components/global/Techs/index.tsx
