+++ src/components/global/Techs/styles.ts
