+++ src/components/global/Typewriter/index.tsx
